refactor(platform): extract export adapters to exportadapters subpackage (#756)
First incremental step decomposing the pkg/platform god-package. The six
Trino/api-gateway export adapters that lived inline in platform.go move to a
self-contained pkg/platform/exportadapters package, collapsed to two exporters
(TrinoExporter, APIExporter). Each implements all three toolkit export
interfaces and takes its portal-store dependencies as constructor parameters,
so wiring is type-checked data flow rather than field access on Platform.
The wire*Export methods stay on Platform and construct one exporter per toolkit.

Reuse portal primitives instead of re-landing duplicated helpers:
- share IDs use uuid.New().String() (was a hand-rolled RFC-4122 generator that
  silently discarded the crypto/rand error)
- share tokens use the newly-exported portal.GenerateShareToken
- the default notice uses the newly-exported portal.DefaultNoticeText

platform.go drops 4491 -> 4254 lines; the moved code leaves the pkg/platform
package entirely. New package coverage 97.7%; make verify green.
